+++ src/ballistica/app/app_globals.h
@@ -56,7 +56,7 @@ class AppGlobals {
   UIScale ui_scale{UIScale::kLarge};
   AccountType account_type{AccountType::kInvalid};
   bool remote_server_accepting_connections{true};
-  std::string game_commands;
+  std::string exec_command;
   std::string user_agent_string{"BA_USER_AGENT_UNSET (" BA_PLATFORM_STRING ")"};
   int return_value{};
   bool is_stdin_a_terminal{true};