+++ src/ballistica/input/device/touch_input.cc
@@ -71,7 +71,7 @@ void TouchInput::HandleTouchEvent(TouchEvent::Type type, void* touch, float x,
 }
 
 TouchInput::TouchInput() {
-  switch (GetInterfaceType()) {
+  switch (GetUIScale()) {
     case UIScale::kSmall:
       base_controls_scale_ = 2.0f;
       world_draw_scale_ = 1.2f;
@@ -866,7 +866,7 @@ void TouchInput::UpdateMapping() {
       g_app_config->Resolve(AppConfig::BoolID::kTouchControlsSwipeHidden);
 
   // Start with defaults.
-  switch (GetInterfaceType()) {
+  switch (GetUIScale()) {
     case UIScale::kSmall:
       buttons_default_frac_x_ = 0.88f;
       buttons_default_frac_y_ = 0.2f;