+++ src/ballistica/ui/widget/root_widget.cc
@@ -141,7 +141,7 @@ auto RootWidget::AddCover(float h_align, VAlign v_align, float x, float y,
   // when the user specifies no backing it means they intend to cover the screen
   // with a flat-ish window texture.. however this only applies to phone-size;
   // for other sizes we always draw a backing.
-  if (GetInterfaceType() != UIScale::kSmall) {
+  if (GetUIScale() != UIScale::kSmall) {
     bd.visibility_mask |=
         static_cast<uint32_t>(Widget::ToolbarVisibility::kMenuFull);
   }
@@ -152,7 +152,7 @@ auto RootWidget::AddCover(float h_align, VAlign v_align, float x, float y,
 
 void RootWidget::AddMeter(float h_align, float x, int type, float r, float g,
                           float b, bool plus, const std::string& s) {
-  float yoffs = (GetInterfaceType() == UIScale::kSmall) ? 0.0f : -7.0f;
+  float yoffs = (GetUIScale() == UIScale::kSmall) ? 0.0f : -7.0f;
 
   float width = type == 1 ? 80.0f : 110.0f;
   // bar
@@ -170,7 +170,7 @@ void RootWidget::AddMeter(float h_align, float x, int type, float r, float g,
     bd.color_r = 0.32f;
     bd.color_g = 0.30f;
     bd.color_b = 0.4f;
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       bd.color_r *= TOOLBAR_COLOR_R;
       bd.color_g *= TOOLBAR_COLOR_G;
       bd.color_b *= TOOLBAR_COLOR_B;
@@ -286,7 +286,7 @@ void RootWidget::AddMeter(float h_align, float x, int type, float r, float g,
     bd.color_r = 0.35f;
     bd.color_g = 0.35f;
     bd.color_b = 0.55f;
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       bd.color_r *= TOOLBAR_COLOR_R;
       bd.color_g *= TOOLBAR_COLOR_G;
       bd.color_b *= TOOLBAR_COLOR_B;
@@ -349,9 +349,9 @@ void RootWidget::Setup() {
 
   // widen this a bit in small mode so it just covers most of the top
   // - that looks funny in medium/large mode though
-  // if (GetInterfaceType() == UIScale::kSmall) {
+  // if (GetUIScale() == UIScale::kSmall) {
   //   AddCover(0.5f, VAlign::kTop, 0.0f, 320.0f,
-  //             GetInterfaceType() == UIScale::kSmall ? 1000.0f :
+  //             GetUIScale() == UIScale::kSmall ? 1000.0f :
   //             1000.0f, 800.0f, 0.4f);
   // }
   // if (c) {
@@ -365,14 +365,14 @@ void RootWidget::Setup() {
     bd.h_align = 0.5f;
     bd.v_align = VAlign::kTop;
     bd.width = 370.0f;
-    // if (GetInterfaceType() != UIScale::kSmall) {
+    // if (GetUIScale() != UIScale::kSmall) {
     //   bd.width = 950.0f;
     // }
     bd.height = 90.0f;
     bd.x = 256.0f;
     bd.y = -20.0f;
     bd.img = "uiAtlas2";
-    // if (GetInterfaceType() != UIScale::kSmall) {
+    // if (GetUIScale() != UIScale::kSmall) {
     //   bd.model_transparent = "toolbarBackingTop";
     // } else {
     bd.model_transparent = "toolbarBackingTop2";
@@ -382,7 +382,7 @@ void RootWidget::Setup() {
     bd.color_g = 0.41f;
     bd.color_b = 0.56f;
     bd.opacity = 1.0f;
-    // if (GetInterfaceType() != UIScale::kSmall) {
+    // if (GetUIScale() != UIScale::kSmall) {
     //   bd.color_r *= TOOLBAR_COLOR_R;
     //   bd.color_g *= TOOLBAR_COLOR_G;
     //   bd.color_b *= TOOLBAR_COLOR_B;
@@ -409,14 +409,14 @@ void RootWidget::Setup() {
     bd.h_align = 0.5f;
     bd.v_align = VAlign::kTop;
     bd.width = 850.0f;
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       bd.width = 850.0f;
     }
     bd.height = 90.0f;
     bd.x = 0.0f;
     bd.y = -20.0f;
     bd.img = "uiAtlas2";
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       bd.model_transparent = "toolbarBackingTop2";
     } else {
       bd.model_transparent = "toolbarBackingTop2";
@@ -426,7 +426,7 @@ void RootWidget::Setup() {
     bd.color_g = 0.41f;
     bd.color_b = 0.56f;
     bd.opacity = 1.0f;
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       bd.color_r *= TOOLBAR_COLOR_R * TOOLBAR_BACK_COLOR_R;
       bd.color_g *= TOOLBAR_COLOR_G * TOOLBAR_BACK_COLOR_G;
       bd.color_b *= TOOLBAR_COLOR_B * TOOLBAR_BACK_COLOR_B;
@@ -446,7 +446,7 @@ void RootWidget::Setup() {
     AddButton(bd);
   }
 
-  float yoffs = (GetInterfaceType() == UIScale::kSmall) ? 0.0f : -10.0f;
+  float yoffs = (GetUIScale() == UIScale::kSmall) ? 0.0f : -10.0f;
 
   // account button
   {
@@ -456,12 +456,12 @@ void RootWidget::Setup() {
     bd.width = 160.0f;
     bd.height = 60.0f;
     bd.depth_min = 0.3f;
-    bd.x = (GetInterfaceType() == UIScale::kSmall) ? 100.0f : -50.0f;
+    bd.x = (GetUIScale() == UIScale::kSmall) ? 100.0f : -50.0f;
     bd.y = -24.0f + yoffs;
     bd.color_r = 0.56f;
     bd.color_g = 0.5f;
     bd.color_b = 0.73f;
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       bd.color_r *= TOOLBAR_COLOR_R;
       bd.color_g *= TOOLBAR_COLOR_G;
       bd.color_b *= TOOLBAR_COLOR_B;
@@ -473,7 +473,7 @@ void RootWidget::Setup() {
          | static_cast<int>(Widget::ToolbarVisibility::kMenuFullRoot));
 
     // on desktop, stick this in the top left corner
-    // if (GetInterfaceType() == UIScale::kLarge) {
+    // if (GetUIScale() == UIScale::kLarge) {
     //   bd.h_align = 0.0f;
     //   bd.x = 120.0f;
     // }
@@ -511,7 +511,7 @@ void RootWidget::Setup() {
     }
   }
 
-  float anchorx = (GetInterfaceType() == UIScale::kSmall) ? 0.3f : 0.25f;
+  float anchorx = (GetUIScale() == UIScale::kSmall) ? 0.3f : 0.25f;
 
   AddMeter(anchorx, 200.0f - 148.0f, 0, 1.0f, 1.0f, 1.0f, false, "456/1000");
   AddMeter(anchorx, 200.0f, 1, 1.0f, 1.0f, 1.0f, false, "123");
@@ -657,7 +657,7 @@ void RootWidget::Setup() {
     float backingCoverG = backingG;
     float backingCoverB = backingB;
     float backingA = 1.0f;
-    if (GetInterfaceType() != UIScale::kSmall) {
+    if (GetUIScale() != UIScale::kSmall) {
       backingR *= TOOLBAR_COLOR_R * TOOLBAR_BACK_COLOR_R;
       backingG *= TOOLBAR_COLOR_G * TOOLBAR_BACK_COLOR_G;
       backingB *= TOOLBAR_COLOR_B * TOOLBAR_BACK_COLOR_B;
@@ -685,7 +685,7 @@ void RootWidget::Setup() {
       bd.x = 0.0f;
       bd.y = 41.0f;
       bd.img = "uiAtlas2";
-      if (GetInterfaceType() != UIScale::kSmall) {
+      if (GetUIScale() != UIScale::kSmall) {
         bd.model_transparent = "toolbarBackingBottom2";
       } else {
         bd.model_transparent = "toolbarBackingBottom2";
@@ -945,7 +945,7 @@ void RootWidget::StepPositions(float dt) {
     }
     if (&b == back_button_) {
       // back button is always disabled in medium/large UI
-      if (GetInterfaceType() != UIScale::kSmall) {
+      if (GetUIScale() != UIScale::kSmall) {
         enable_button = false;
       }
 
@@ -1034,7 +1034,7 @@ void RootWidget::StepPositions(float dt) {
 void RootWidget::UpdateLayout() {
   // Now actually put things in place.
   base_scale_ = 1.0f;
-  switch (GetInterfaceType()) {
+  switch (GetUIScale()) {
     case UIScale::kLarge:
       base_scale_ = 0.6f;
       break;